requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pyjwt>=2.8.0
httpx>=0.24.1
pydantic>=2.3.0
python-dotenv>=1.0.0